renamed _types.py as _typing.py
