+++ skbio/util/_typing.py
